# Perf rearchitecting