--- a/src/components/DayPickerRangeController.jsx
+++ b/src/components/DayPickerRangeController.jsx
@@ -12,6 +12,14 @@ import isTouchDevice from '../utils/isTouchDevice';
 import isInclusivelyAfterDay from '../utils/isInclusivelyAfterDay';
 import isNextDay from '../utils/isNextDay';
 import isSameDay from '../utils/isSameDay';
+import isAfterDay from '../utils/isAfterDay';
+import isBeforeDay from '../utils/isBeforeDay';
+
+import getVisibleDays from '../utils/getVisibleDays';
+import isDayVisible from '../utils/isDayVisible';
+
+import toISODateString from '../utils/toISODateString';
+import toISOMonthString from '../utils/toISOMonthString';
 
 import FocusedInputShape from '../shapes/FocusedInputShape';
 import ScrollableOrientationShape from '../shapes/ScrollableOrientationShape';
@@ -121,24 +129,184 @@ const defaultProps = {
 export default class DayPickerRangeController extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {
-      hoverDate: null,
-      phrases: props.phrases,
-    };
 
     this.isTouchDevice = isTouchDevice();
     this.today = moment();
+    this.modifiers = {
+      today: day => this.isToday(day),
+      blocked: day => this.isBlocked(day),
+      'blocked-calendar': day => props.isDayBlocked(day),
+      'blocked-out-of-range': day => props.isOutsideRange(day),
+      'highlighted-calendar': day => props.isDayHighlighted(day),
+      valid: day => !this.isBlocked(day),
+      'selected-start': day => this.isStartDate(day),
+      'selected-end': day => this.isEndDate(day),
+      'blocked-minimum-nights': day => this.doesNotMeetMinimumNights(day),
+      'selected-span': day => this.isInSelectedSpan(day),
+      'last-in-range': day => this.isLastInRange(day),
+      hovered: day => this.isHovered(day),
+      'hovered-span': day => this.isInHoveredSpan(day),
+      'after-hovered-start': day => this.isDayAfterHoveredStartDate(day),
+    };
+
+    const { currentMonth, visibleDays } = this.getStateForNewMonth(props);
+
+    this.state = {
+      hoverDate: null,
+      currentMonth,
+      phrases: props.phrases,
+      visibleDays,
+    };
 
     this.onDayClick = this.onDayClick.bind(this);
     this.onDayMouseEnter = this.onDayMouseEnter.bind(this);
     this.onDayMouseLeave = this.onDayMouseLeave.bind(this);
+    this.onPrevMonthClick = this.onPrevMonthClick.bind(this);
+    this.onNextMonthClick = this.onNextMonthClick.bind(this);
     this.getFirstFocusableDay = this.getFirstFocusableDay.bind(this);
   }
 
   componentWillReceiveProps(nextProps) {
-    const { focusedInput, phrases } = nextProps;
+    const {
+      startDate,
+      endDate,
+      focusedInput,
+      minimumNights,
+      isOutsideRange,
+      isDayBlocked,
+      isDayHighlighted,
+      phrases,
+      initialVisibleMonth,
+      numberOfMonths,
+      enableOutsideDays,
+    } = nextProps;
+    let { visibleDays } = this.state;
 
-    if (focusedInput !== this.props.focusedInput || phrases !== this.props.phrases) {
+    if (isOutsideRange !== this.props.isOutsideRange) {
+      this.modifiers['blocked-out-of-range'] = day => isOutsideRange(day);
+    }
+
+    if (isDayBlocked !== this.props.isDayBlocked) {
+      this.modifiers['blocked-calendar'] = day => isDayBlocked(day);
+    }
+
+    if (isDayHighlighted !== this.props.isDayHighlighted) {
+      this.modifiers['highlighted-calendar'] = day => isDayHighlighted(day);
+    }
+
+    if (
+      initialVisibleMonth !== this.props.initialVisibleMonth ||
+      numberOfMonths !== this.props.numberOfMonths ||
+      enableOutsideDays !== this.props.enableOutsideDays
+    ) {
+      const newMonthState = this.getStateForNewMonth(nextProps);
+      const currentMonth = newMonthState.currentMonth;
+      visibleDays = newMonthState.visibleDays;
+      this.setState({
+        currentMonth,
+        visibleDays,
+      });
+    }
+
+    const didStartDateChange = startDate !== this.props.startDate;
+    const didEndDateChange = endDate !== this.props.endDate;
+    const didFocusChange = focusedInput !== this.props.focusedInput;
+
+    let modifiers = {};
+
+    if (didStartDateChange) {
+      modifiers = this.deleteModifier(modifiers, this.props.startDate, 'selected-start');
+      modifiers = this.addModifier(modifiers, startDate, 'selected-start');
+    }
+
+    if (didEndDateChange) {
+      modifiers = this.deleteModifier(modifiers, this.props.endDate, 'selected-end');
+      modifiers = this.addModifier(modifiers, endDate, 'selected-end');
+    }
+
+    if (didStartDateChange || didEndDateChange) {
+      if (this.props.startDate && this.props.endDate) {
+        modifiers = this.deleteModifierFromRange(
+          modifiers,
+          this.props.startDate.clone().add(1, 'day'),
+          this.props.endDate,
+          'selected-span',
+        );
+      }
+
+      if (startDate && endDate) {
+        modifiers = this.deleteModifierFromRange(
+          modifiers,
+          startDate,
+          endDate,
+          'hovered-span',
+        );
+
+        modifiers = this.addModifierToRange(
+          modifiers,
+          startDate.clone().add(1, 'day'),
+          endDate,
+          'selected-span',
+        );
+      }
+    }
+
+    if (minimumNights > 0 || minimumNights !== this.props.minimumNights) {
+      if (this.props.startDate && (didFocusChange || didStartDateChange)) {
+        modifiers = this.deleteModifierFromRange(
+          modifiers,
+          this.props.startDate,
+          this.props.startDate.clone().add(minimumNights, 'days'),
+          'blocked-minimum-nights',
+        );
+      }
+
+      if (startDate && focusedInput === END_DATE) {
+        modifiers = this.addModifierToRange(
+          modifiers,
+          startDate,
+          startDate.clone().add(minimumNights, 'days'),
+          'blocked-minimum-nights',
+        );
+      }
+    }
+
+    if (didFocusChange) {
+      Object.values(visibleDays).forEach((days) => {
+        Object.keys(days).forEach((day) => {
+          const momentObj = moment(day);
+          if (isDayBlocked(momentObj)) {
+            modifiers = this.addModifier(modifiers, momentObj, 'blocked-calendar');
+          } else {
+            modifiers = this.deleteModifier(modifiers, momentObj, 'blocked-calendar');
+          }
+
+          if (isDayHighlighted(momentObj)) {
+            modifiers = this.addModifier(modifiers, momentObj, 'highlighted-calendar');
+          } else {
+            modifiers = this.deleteModifier(modifiers, momentObj, 'highlighted-calendar');
+          }
+        });
+      });
+    }
+
+    const today = moment();
+    if (!isSameDay(this.today, today)) {
+      modifiers = this.deleteModifier(modifiers, this.today, 'today');
+      modifiers = this.addModifier(modifiers, today, 'today');
+      this.today = today;
+    }
+
+    if (Object.keys(modifiers).length > 0) {
+      this.setState({
+        visibleDays: {
+          ...visibleDays,
+          ...modifiers,
+        },
+      });
+    }
+
+    if (didFocusChange || phrases !== this.props.phrases) {
       // set the appropriate CalendarDay phrase based on focusedInput
       let chooseAvailableDate = phrases.chooseAvailableDate;
       if (focusedInput === START_DATE) {
@@ -154,10 +322,6 @@ export default class DayPickerRangeController extends React.Component {
         },
       });
     }
-  }
-
-  componentWillUpdate() {
-    this.today = moment();
   }
 
   onDayClick(day, e) {
@@ -200,18 +364,111 @@ export default class DayPickerRangeController extends React.Component {
 
   onDayMouseEnter(day) {
     if (this.isTouchDevice) return;
+    const { startDate, endDate, focusedInput } = this.props;
+    const { hoverDate, visibleDays } = this.state;
+
+    let modifiers = {};
+    modifiers = this.deleteModifier(modifiers, hoverDate, 'hovered');
+    modifiers = this.addModifier(modifiers, day, 'hovered');
+
+    if (startDate && !endDate && focusedInput === END_DATE) {
+      if (isAfterDay(hoverDate, startDate)) {
+        modifiers = this.deleteModifierFromRange(modifiers, startDate, hoverDate, 'hovered-span');
+      }
+
+      if (!this.isBlocked(day) && isAfterDay(day, startDate)) {
+        modifiers = this.addModifierToRange(modifiers, startDate, day, 'hovered-span');
+      }
+    }
+
+    if (!startDate && endDate && focusedInput === START_DATE) {
+      if (isBeforeDay(hoverDate, endDate)) {
+        modifiers = this.deleteModifierFromRange(modifiers, hoverDate, endDate, 'hovered-span');
+      }
+
+      if (!this.isBlocked(day) && isBeforeDay(day, endDate)) {
+        modifiers = this.addModifierToRange(modifiers, day, endDate, 'hovered-span');
+      }
+    }
 
     this.setState({
       hoverDate: day,
+      visibleDays: {
+        ...visibleDays,
+        ...modifiers,
+      },
     });
   }
 
   onDayMouseLeave() {
-    if (this.isTouchDevice) return;
+    const { startDate, endDate } = this.props;
+    const { hoverDate, visibleDays } = this.state;
+    if (this.isTouchDevice || !hoverDate) return;
+
+    let modifiers = {};
+    modifiers = this.deleteModifier(modifiers, hoverDate, 'hovered');
+
+    if (startDate && !endDate && isAfterDay(hoverDate, startDate)) {
+      modifiers = this.deleteModifierFromRange(modifiers, startDate, hoverDate, 'hovered-span');
+    }
+
+    if (!startDate && endDate && isAfterDay(endDate, hoverDate)) {
+      modifiers = this.deleteModifierFromRange(modifiers, hoverDate, endDate, 'hovered-span');
+    }
 
     this.setState({
       hoverDate: null,
+      visibleDays: {
+        ...visibleDays,
+        ...modifiers,
+      },
     });
+  }
+
+  onPrevMonthClick() {
+    const { onPrevMonthClick, numberOfMonths, enableOutsideDays } = this.props;
+    const { currentMonth, visibleDays } = this.state;
+
+    const newVisibleDays = {};
+    Object.keys(visibleDays).sort().slice(0, numberOfMonths + 1).forEach((month) => {
+      newVisibleDays[month] = visibleDays[month];
+    });
+
+    const prevMonth = currentMonth.clone().subtract(1, 'months');
+    const prevMonthVisibleDays = getVisibleDays(prevMonth, 1, enableOutsideDays);
+
+    this.setState({
+      currentMonth: prevMonth,
+      visibleDays: {
+        ...newVisibleDays,
+        ...this.getModifiers(prevMonthVisibleDays),
+      },
+    });
+
+    onPrevMonthClick();
+  }
+
+  onNextMonthClick() {
+    const { onNextMonthClick, numberOfMonths, enableOutsideDays } = this.props;
+    const { currentMonth, visibleDays } = this.state;
+
+    const newVisibleDays = {};
+    Object.keys(visibleDays).sort().slice(1).forEach((month) => {
+      newVisibleDays[month] = visibleDays[month];
+    });
+
+    const nextMonth = currentMonth.clone().add(numberOfMonths, 'month');
+    const nextMonthVisibleDays = getVisibleDays(nextMonth, 1, enableOutsideDays);
+
+    this.setState({
+      currentMonth: currentMonth.clone().add(1, 'month'),
+      visibleDays: {
+        ...newVisibleDays,
+        ...this.getModifiers(nextMonthVisibleDays),
+      },
+    });
+
+    onNextMonthClick();
   }
 
   getFirstFocusableDay(newMonth) {
@@ -230,16 +487,135 @@ export default class DayPickerRangeController extends React.Component {
       const days = [];
       const lastVisibleDay = newMonth.clone().add(numberOfMonths - 1, 'months').endOf('month');
       let currentDay = focusedDate.clone();
-      while (!currentDay.isAfter(lastVisibleDay)) {
+      while (!isAfterDay(currentDay, lastVisibleDay)) {
         currentDay = currentDay.clone().add(1, 'day');
         days.push(currentDay);
       }
 
-      const viableDays = days.filter(day => !this.isBlocked(day) && day.isAfter(focusedDate));
+      const viableDays = days.filter(day => !this.isBlocked(day));
+
       if (viableDays.length > 0) focusedDate = viableDays[0];
     }
 
     return focusedDate;
+  }
+
+  getModifiers(visibleDays) {
+    const modifiers = {};
+    Object.keys(visibleDays).forEach((month) => {
+      modifiers[month] = {};
+      visibleDays[month].forEach((day) => {
+        modifiers[month][toISODateString(day)] = this.getModifiersForDay(day);
+      });
+    });
+
+    return modifiers;
+  }
+
+  getModifiersForDay(day) {
+    return new Set(Object.keys(this.modifiers).filter(modifier => this.modifiers[modifier](day)));
+  }
+
+  getStateForNewMonth(nextProps) {
+    const { initialVisibleMonth, numberOfMonths, enableOutsideDays } = nextProps;
+    const currentMonth = initialVisibleMonth();
+    const visibleDays =
+      this.getModifiers(getVisibleDays(currentMonth, numberOfMonths, enableOutsideDays));
+    return { currentMonth, visibleDays };
+  }
+
+  addModifier(updatedDays, day, modifier) {
+    const { numberOfMonths, enableOutsideDays } = this.props;
+    const { currentMonth, visibleDays } = this.state;
+    if (!day || !isDayVisible(day, currentMonth, numberOfMonths, enableOutsideDays)) {
+      return updatedDays;
+    }
+
+    let monthIso = toISOMonthString(day);
+    let month = updatedDays[monthIso] || visibleDays[monthIso];
+    const iso = toISODateString(day);
+    if (enableOutsideDays) {
+      const startOfMonth = day.clone().startOf('month');
+      const endOfMonth = day.clone().endOf('month');
+      if (
+        isBeforeDay(startOfMonth, currentMonth.clone().startOf('month')) ||
+        isAfterDay(endOfMonth, currentMonth.clone().endOf('month'))
+      ) {
+        monthIso = Object.keys(visibleDays).filter(monthKey => (
+          monthKey !== monthIso && Object.keys(visibleDays[monthKey]).indexOf(iso) > -1
+        ))[0];
+        month = updatedDays[monthIso] || visibleDays[monthIso];
+      }
+    }
+
+    const modifiers = new Set(month[iso]);
+    modifiers.add(modifier);
+    return {
+      ...updatedDays,
+      [monthIso]: {
+        ...month,
+        [iso]: modifiers,
+      },
+    };
+  }
+
+  addModifierToRange(updatedDays, start, end, modifier) {
+    let days = updatedDays;
+
+    let spanStart = start.clone();
+    while (isBeforeDay(spanStart, end)) {
+      days = this.addModifier(days, spanStart, modifier);
+      spanStart = spanStart.clone().add(1, 'day');
+    }
+
+    return days;
+  }
+
+  deleteModifier(updatedDays, day, modifier) {
+    const { numberOfMonths, enableOutsideDays } = this.props;
+    const { currentMonth, visibleDays } = this.state;
+    if (!day || !isDayVisible(day, currentMonth, numberOfMonths, enableOutsideDays)) {
+      return updatedDays;
+    }
+
+    let monthIso = toISOMonthString(day);
+    let month = updatedDays[monthIso] || visibleDays[monthIso];
+    const iso = toISODateString(day);
+    if (enableOutsideDays) {
+      const startOfMonth = day.clone().startOf('month');
+      const endOfMonth = day.clone().endOf('month');
+      if (
+        isBeforeDay(startOfMonth, currentMonth.clone().startOf('month')) ||
+        isAfterDay(endOfMonth, currentMonth.clone().endOf('month'))
+      ) {
+        monthIso = Object.keys(visibleDays).filter(monthKey => (
+          monthKey !== monthIso && Object.keys(visibleDays[monthKey]).indexOf(iso) > -1
+        ))[0];
+        month = updatedDays[monthIso] || visibleDays[monthIso];
+      }
+    }
+
+    const modifiers = new Set(month[iso]);
+    modifiers.delete(modifier);
+    return {
+      ...updatedDays,
+      [monthIso]: {
+        ...month,
+        [iso]: modifiers,
+      },
+    };
+  }
+
+  deleteModifierFromRange(updatedDays, start, end, modifier) {
+    let days = updatedDays;
+
+    let spanStart = start.clone();
+    while (isBeforeDay(spanStart, end)) {
+      days = this.deleteModifier(days, spanStart, modifier);
+      spanStart = spanStart.clone().add(1, 'day');
+    }
+
+    return days;
   }
 
   doesNotMeetMinimumNights(day) {
@@ -255,7 +631,7 @@ export default class DayPickerRangeController extends React.Component {
 
   isDayAfterHoveredStartDate(day) {
     const { startDate, endDate, minimumNights } = this.props;
-    const { hoverDate } = this.state;
+    const { hoverDate } = this.state || {};
     return !!startDate && !endDate && !this.isBlocked(day) && isNextDay(hoverDate, day) &&
       minimumNights > 0 && isSameDay(hoverDate, startDate);
   }
@@ -265,12 +641,13 @@ export default class DayPickerRangeController extends React.Component {
   }
 
   isHovered(day) {
-    return isSameDay(day, this.state.hoverDate);
+    const { hoverDate } = this.state || {};
+    return isSameDay(day, hoverDate);
   }
 
   isInHoveredSpan(day) {
     const { startDate, endDate } = this.props;
-    const { hoverDate } = this.state;
+    const { hoverDate } = this.state || {};
 
     const isForwardRange = !!startDate && !endDate &&
       (day.isBetween(startDate, hoverDate) ||
@@ -308,9 +685,6 @@ export default class DayPickerRangeController extends React.Component {
 
   render() {
     const {
-      isDayBlocked,
-      isDayHighlighted,
-      isOutsideRange,
       numberOfMonths,
       orientation,
       monthFormat,
@@ -318,8 +692,6 @@ export default class DayPickerRangeController extends React.Component {
       navPrev,
       navNext,
       onOutsideClick,
-      onPrevMonthClick,
-      onNextMonthClick,
       withPortal,
       enableOutsideDays,
       initialVisibleMonth,
@@ -328,59 +700,26 @@ export default class DayPickerRangeController extends React.Component {
       focusedInput,
       renderDay,
       renderCalendarInfo,
-      startDate,
-      endDate,
       onBlur,
       isFocused,
       showKeyboardShortcuts,
       isRTL,
     } = this.props;
 
-    const { phrases } = this.state;
-
-    const modifiers = {
-      today: day => this.isToday(day),
-      blocked: day => this.isBlocked(day),
-      'blocked-calendar': day => isDayBlocked(day),
-      'blocked-out-of-range': day => isOutsideRange(day),
-      'highlighted-calendar': day => isDayHighlighted(day),
-      valid: day => !this.isBlocked(day),
-
-      // Modifiers are computed for every CalendarDay, so we omit where
-      // logically possible.
-      ...startDate && {
-        'selected-start': day => this.isStartDate(day),
-      },
-      ...endDate && {
-        'selected-end': day => this.isEndDate(day),
-        'blocked-minimum-nights': day => this.doesNotMeetMinimumNights(day),
-      },
-      ...(startDate && endDate) && {
-        'selected-span': day => this.isInSelectedSpan(day),
-        'last-in-range': day => this.isLastInRange(day),
-      },
-      ...!this.isTouchDevice && {
-        // before anything has been set or after both are set
-        hovered: day => this.isHovered(day),
-
-        // while start date has been set, but end date has not been
-        'hovered-span': day => this.isInHoveredSpan(day),
-        'after-hovered-start': day => this.isDayAfterHoveredStartDate(day),
-      },
-    };
+    const { phrases, visibleDays } = this.state;
 
     return (
       <DayPicker
         ref={(ref) => { this.dayPicker = ref; }}
         orientation={orientation}
         enableOutsideDays={enableOutsideDays}
-        modifiers={modifiers}
+        modifiers={visibleDays}
         numberOfMonths={numberOfMonths}
         onDayClick={this.onDayClick}
         onDayMouseEnter={this.onDayMouseEnter}
         onDayMouseLeave={this.onDayMouseLeave}
-        onPrevMonthClick={onPrevMonthClick}
-        onNextMonthClick={onNextMonthClick}
+        onPrevMonthClick={this.onPrevMonthClick}
+        onNextMonthClick={this.onNextMonthClick}
         monthFormat={monthFormat}
         renderMonth={renderMonth}
         withPortal={withPortal}