This PR is a dramatic rearchitecture of the way that we handle modifiers in `react-dates` on the whole.

Instead of passing down an object of modifier string => modifiers functions to each CalendarDay component (one that gets rebuilt on every render call) and expecting the CalendarDay component to take care of updating itself, we now put the burden on the top of the tree instead of on the leaves. The previous model meant that whenever an interaction happened on an individual day, even a hover interaction, every single CalendarDay would have to recalculate its modifiers and rerender as a result. The idea was that an interaction on one day might have effects elsewhere in the calendar (like the `hovered-span`). The effect unfortunately was that react-dates was [really god damn slow](https://github.com/airbnb/react-dates/issues/272).

In this new model, the DayPickerRangeController maintains a map with the following structure:
```
{
  MONTH_ISO_1: {
    DAY_ISO_1: Set(['modifier_1', 'modifier_2', ...]),
    DAY_ISO_2: Set(['modifier_1', 'modifier_2', ...]),
    ...
  },
  ...
}
```
It passes this down the tree such that each `CalendarMonth` and each `CalendarDay` only gets the information that pertains to it. This means that the updating of these modifiers is also handled at the top-level and is done in the `componentWillReceiveProps`, `onDayMouseEnter`, and `onDayMouseLeave` methods of the `DayPickerRangeController`. Fortunately, this allows us to more finely tune which days get updated and speeds up the rerendering/updating process dramatically.

I did some preliminary testing and saw the following improvements:
*Before:*
<img width="1100" alt="screen shot 2017-04-13 at 10 35 39 pm" src="https://cloud.githubusercontent.com/assets/1383861/25154213/9d8ba8aa-2444-11e7-847a-1762f164ac6e.png">

*After:*
<img width="1382" alt="screen shot 2017-04-13 at 10 55 41 pm" src="https://cloud.githubusercontent.com/assets/1383861/25154217/a688b0c4-2444-11e7-8214-58b2904eccc3.png">

I will update those graphs when everything is in place.

I still need to finish writing some tests, clean up some other tests, and also get this working with `enableOutsideDays` and the `SingleDatePicker`.  The bulk of the work is in `DayPickerRangeController` right now, but I've been musing on ways to pull that in an HOC or something. SUGGESTIONS ARE WELCOME.

In any case, I would appreciate a first review pass while I finish up the above points! <3
@lencioni @ljharb @moonboots @backwardok 

UPDATE: Here are the final perf measurements!
**Before:**
<img width="1270" alt="screen shot 2017-05-01 at 3 19 09 pm" src="https://cloud.githubusercontent.com/assets/1383861/25598878/97062454-2e8c-11e7-81d6-45555caa6652.png">

**After:**
![screen shot 2017-05-01 at 4 38 44 pm](https://cloud.githubusercontent.com/assets/1383861/25598893/accb5728-2e8c-11e7-94fe-dd52b5507e3f.png)

Here are the findings:
 - The overhead on initial mount remains about the same (~400-500ms)
 - Any calendar interaction (hover, click, etc.) consistently caused a 65ms update in the old version and now only takes about 10ms - 15ms.
 - Transitions are now a bit slower (likely due to a bug in the code and therefore an opportunity) at about 120ms as opposed to 95ms previously. There does seem to be an unnecessary update occurring that I'll investigate in a follow up PR.