@majapw 2017-05-01
@ljharb I think I addressed your concerns! Can you take a look?