@dburles 2017-06-14
Hey guys, I have the following on a `DayPicker` component in order to highlight a range of dates that contain information, along with highlighting the currently selected day. However it now no longer works after this refactor. Is there another approach I can take to make it work? thanks.

```
            modifiers={{
              selected: day =>
                moment(this.props.selectedDay).isSame(day, 'day'),
              'highlighted-calendar': day1 =>
                this.props.highlightedDays.some(day2 =>
                  moment(day2).isSame(day1, 'day'),
                ),
            }}
```