@majapw 2017-06-14
Hey @dburles, can you open an issue for this so that we can have a convo on a separate thread?